Add stacking algorithms, improve alignment, fix time estimator
Algorithms:
- Weighted Average: proper accumulation (no pairwise blur)
- Depth Map: multi-res sharpness with edge-aware bilateral smoothing
- Exposure Fusion (HDR): batched Mertens, clearly labeled for HDR not focus
- Laplacian: contrast threshold, feathered focusmaps, local tone-mapping
- Configurable DMap smoothing slider (Zerene-style quality control)
- Auto-detect parameters from image resolution on import

Alignment:
- Multi-scale ECC (512->1024->2048) for precise sub-pixel alignment
- MOTION_EUCLIDEAN (3 DOF) instead of AFFINE (6 DOF) - correct model
- Auto-crop works with rotation+scale alignment

UX:
- Time estimator rewritten: median-based, skips JIT warmup, no fudge factor
- Settings locked during stacking, unlocked on cancel/finish
- Auto-preview stacked result, clear stale selection
- Hidden/system files auto-filtered on folder import
- Unsupported files show status bar message instead of popup
- Advanced settings collapsible with edge-aware controls

CLI:
- Progress bar with ETA and speed
- --auto, --auto-crop, --method, --rotation-scale flags

59 tests passing.

+++ src/algorithms/dft_imreg.py
@@ -958,7 +958,7 @@ class im_reg:
         # Store last computed shift for external access
         self.last_shift = (float(x_shift), float(y_shift))
 
-        return result.astype(np.uint8)
+        return result.astype(np.float32)
 
     # TODO: Implement with option of selecting what method to use (not yet successfully implemented)
     # Register im1 to im0 for Rotation, Scale, Translation (RST)
@@ -1010,4 +1010,4 @@ class im_reg:
         result = cv2.warpAffine(result, rotation, (width, height))
         result = cv2.warpAffine(result, translation, (width, height))
 
-        return result.astype(np.uint8)
+        return result.astype(np.float32)
